# RPC: Add reserve member function to `UniValue` and use it in `getblock` RPC