--- a/src/univalue/include/univalue.h
+++ b/src/univalue/include/univalue.h
@@ -70,6 +70,8 @@ public:
 
     size_t size() const { return values.size(); }
 
+    void reserve(size_t new_cap);
+
     void getObjMap(std::map<std::string,UniValue>& kv) const;
     bool checkObject(const std::map<std::string,UniValue::VType>& memberTypes) const;
     const UniValue& operator[](const std::string& key) const;