@andrewtoth 2024-10-29
I think we can take this further and reserve for VOBJ types instead of just VARR.
The VOBJ uses both values and keys, so we can reserve both. We can count how many times we do `pushKV` and reserve that amount for both keys and values.