@andrewtoth 2024-10-30
I believe you can get more fine grained benchmarks for this using the rpc_blockchain benchmarks.