@ismaelsadeeq 2024-10-30
> I think we can take this further and reserve for `VOBJ` types instead of just `VARR`. The `VOBJ` type uses both values and keys, so we can reserve for both.

Done in the latest push, see [diff](https://github.com/bitcoin/bitcoin/compare/fa73b9a9984d5dd2efa6fd5dd70b3db617803a25..f8d5d2cfc63bea5b58b08ceaa743c26820c1763d), but I am reserving keys only when `typ` is of `VOBJ`.

> We can count how many times we use `pushKV` and reserve that amount for both keys and values.

+1, this approach is possible but is it maintainable?
Manually counting each `pushKV`s means updating the value thats being passed to `.reserve` every time we add another element, will be better to reserve for size's that are determined at runtime or at least use a constant.

Is there a constant for the number of items in a block so that I can reserve for results in `blockheaderToJSON`?