@Eunovo 2024-11-17
> > I believe you can get more fine-grained benchmarks for this using the `rpc_blockchain` benchmarks.
> 
> Yes, I can, but I’ll need to add verbosity levels 1 and 2, where we can see some performance improvements.
> 
> Show diff
> Benchmark Results
> 
> On **master**:
> 
> ns/op	op/s	err%	total	benchmark
> 190,342	5,254	2.3%	0.01	`BlockToJsonVerbose1`
> 34,812,292	28.73	1.0%	0.38	`BlockToJsonVerbose2`
> 34,457,167	29.02	1.0%	0.38	`BlockToJsonVerbose3`
> On **this PR**:
> 
> ns/op	op/s	err%	total	benchmark
> 172,278	5,805	0.7%	0.01	`BlockToJsonVerbose1`
> 33,720,584	29.66	0.4%	0.37	`BlockToJsonVerbose2`
> 33,884,417	29.51	1.2%	0.38	`BlockToJsonVerbose3`

@ismaelsadeeq, could you commit the diff you’ve provided and include the benchmark results in the commit message? Then, rebase your current changes on top, rerun the benchmarks for https://github.com/bitcoin/bitcoin/pull/31179/commits/952463ccdc2baeacb527c66732fd184cbcbd53ba and https://github.com/bitcoin/bitcoin/pull/31179/commits/28e3392d11355b1160dc1a7a5557081728a02840, and update their commit messages with the results.

This way I can just go through your commits, run the benchmarks and compare with your results.