@l0rinc 2024-12-27
@ismaelsadeeq, I ran your benchmarks from https://github.com/bitcoin/bitcoin/pull/31179#issuecomment-2446067820, but I only see a 7% speedup in the `TxVerbosity::SHOW_TXID` bench - while the rest were exactly the same as before.
But if it's *that* important here, we could easily get a 30% speedup as well there by deduplicating the `GetSerializeSize` calls:
```C++
static int64_t GetBlockWeight(const size_t size_no_witness, const size_t size_with_witness)
{
    return size_no_witness * (WITNESS_SCALE_FACTOR - 1) + size_with_witness;
}
static int64_t GetBlockWeight(const CBlock& block)
{
    const size_t size_no_witness = ::GetSerializeSize(TX_NO_WITNESS(block));
    const size_t size_with_witness = ::GetSerializeSize(TX_WITH_WITNESS(block));
    return GetBlockWeight(size_no_witness, size_with_witness);
}
```
...
```C++
UniValue blockToJSON(BlockManager& blockman, const CBlock& block, const CBlockIndex& tip, const CBlockIndex& blockindex, TxVerbosity verbosity)
{
    UniValue result = blockheaderToJSON(tip, blockindex);

    const size_t size_no_witness = GetSerializeSize(TX_NO_WITNESS(block));
    const size_t size_with_witness = GetSerializeSize(TX_WITH_WITNESS(block));
    result.pushKV("strippedsize", (int)size_no_witness);
    result.pushKV("size", (int)size_with_witness);
    result.pushKV("weight", (int)GetBlockWeight(size_no_witness, size_with_witness));
    UniValue txs(UniValue::VARR);
    txs.reserve(block.vtx.size());
...
```

----

Or if you need `TxVerbosity::SHOW_DETAILS` and `TxVerbosity::SHOW_DETAILS_AND_PREVOUT`, they can be sped up by ~20% by reviving the abandoned (for lack of interest) https://github.com/bitcoin/bitcoin/pull/30035.

---

And if `BlockToJsonVerboseWrite` needs a boost, https://github.com/bitcoin/bitcoin/pull/31144 will speed up by ~14%.