@ismaelsadeeq 2024-12-30
Thanks for taking a look @Eunovo @l0rinc 

> could you commit the diff you’ve provided and include the benchmark results in the commit message?

I've added 8d4823b383be4f8cd0152176a4f67b447d29549 For the benchmarks and instead update the PR title with the bench results.

> I ran your benchmarks from https://github.com/bitcoin/bitcoin/pull/31179#issuecomment-2446067820, but I only see a 7% speedup in the TxVerbosity::SHOW_TXID bench - while the rest were exactly the same as before.

I did that locally and the result is the same as well

>  But if it's that important, we could easily get a 30% speedup as well there by deduplicating the GetSerializeSize calls:

I'll be happy to review that as a PR or might attempt it myself.

> Or if you need TxVerbosity::SHOW_DETAILS and TxVerbosity::SHOW_DETAILS_AND_PREVOUT, they can be sped up by ~20% by reviving the abandoned (for lack of interest) https://github.com/bitcoin/bitcoin/pull/30035.

Interesting I will take a look at that, it seems there are some unaddressed comments by @TheCharlatan at that PR

> And if BlockToJsonVerboseWrite needs a boost, https://github.com/bitcoin/bitcoin/pull/31144 will speed up by ~14%.

I will take a look.

> @martinus also contains a few optimization attempts (a lot of them were already applied in other PRs since, but a few of the changes can still be done).
If RPC speed is important to you, the above suggestions can result in a measurable improvements.

RPC speed is important not just to me but to all users. For now, I've been experimenting with using libbitcoinkernel, and it seems to be more performant than the RPC.

I'll look into it further.

All those PRs offer improvements that are beneficial to users. I also think having this reservation ability is important to avoid the reallocation overhead. @l0rinc?