@l0rinc 2024-12-30
I can implement my suggestions in separate PRs if you say the RPC speedups would be welcome